+++ R/class-workbook.R
@@ -319,6 +319,7 @@ wbWorkbook <- R6::R6Class(
     #' Add worksheet to the `wbWorkbook` object
     #' @param sheet sheet
     #' @param gridLines gridLines
+    #' @param rowColHeaders rowColHeaders
     #' @param tabColour tabColour
     #' @param zoom zoom
     #' @param header header
@@ -339,6 +340,7 @@ wbWorkbook <- R6::R6Class(
     add_worksheet = function(
       sheet       = next_sheet(),
       gridLines   = TRUE,
+      rowColHeaders = TRUE,
       tabColour   = NULL,
       zoom        = 100,
       header      = NULL,
@@ -476,6 +478,7 @@ wbWorkbook <- R6::R6Class(
       self$append("worksheets",
         wbWorksheet$new(
           gridLines   = gridLines,
+          rowColHeaders = rowColHeaders,
           tabSelected = newSheetIndex == 1,
           tabColour   = tabColour,
           zoom        = zoom,
@@ -670,13 +673,14 @@ wbWorkbook <- R6::R6Class(
         tbls$tab_name <- stri_join(tbls$tab_name, "_n")
         tbls$tab_sheet <- newSheetIndex
         # modify tab_xml with updated name, displayName and id
-        tbls$tab_xml <- vapply(seq_len(nrow(tbls)), function(x)
+        tbls$tab_xml <- vapply(seq_len(nrow(tbls)), function(x) {
           xml_attr_mod(tbls$tab_xml[x],
                        xml_attributes = c(name = tbls$tab_name[x],
                                           displayName = tbls$tab_name[x],
                                           id = newid[x])
-          ),
-          NA_character_
+          )
+        },
+        NA_character_
         )
 
         # add new tables to old tables
@@ -4108,7 +4112,493 @@ wbWorkbook <- R6::R6Class(
       self$worksheets[[sheet]]$clean_sheet(numbers = numbers, characters = characters, styles = styles, merged_cells = merged_cells)
 
       invisible(self)
+    },
+
+    #' @description create borders for cell region
+    #' @param sheet a worksheet
+    #' @param dims dimensions on the worksheet e.g. "A1", "A1:A5", "A1:H5"
+    #' @param bottom_color,left_color,right_color,top_color,inner_hcolor,inner_vcolor a color, either something openxml knows or some RGB color
+    #' @param left_border,right_border,top_border,bottom_border,inner_hgrid,inner_vgrid the border style, if NULL no border is drawn. See create_border for possible border styles
+    #' @seealso create_border
+    #' @examples
+    #'
+    #' wb <- wb_workbook()
+    #' wb$add_worksheet("S1")$add_data("S1", mtcars)
+    #' wb$add_border(1, dims = "A1:K1",
+    #'  left_border = NULL, right_border = NULL,
+    #'  top_border = NULL, bottom_border = "double")
+    #' wb$add_border(1, dims = "A5",
+    #'  left_border = "dotted", right_border = "dotted",
+    #'  top_border = "hair", bottom_border = "thick")
+    #' wb$add_border(1, dims = "C2:C5")
+    #' wb$add_border(1, dims = "G2:H3")
+    #' wb$add_border(1, dims = "G12:H13",
+    #'  left_color = c(rgb = "FF9400D3"), right_color = c(rgb = "FF4B0082"),
+    #'  top_color = c(rgb = "FF0000FF"), bottom_color = c(rgb = "FF00FF00"))
+    #' wb$add_border(1, dims = "A20:C23")
+    #' wb$add_border(1, dims = "B12:D14",
+    #'  left_color = c(rgb = "FFFFFF00"), right_color = c(rgb = "FFFF7F00"),
+    #'  bottom_color = c(rgb ="FFFF0000"))
+    #' wb$add_border(1, dims = "D28:E28")
+    #' # if (interactive()) wb$open()
+    #'
+    #' wb <- wb_workbook()
+    #' wb$add_worksheet("S1")$add_data("S1", mtcars)
+    #' wb$add_border(1, dims = "A2:K33", inner_vgrid = "thin", inner_vcolor = c(rgb="FF808080"))
+    #' @return The `wbWorksheetObject`, invisibly
+    add_border = function(
+      sheet         = 1,
+      dims          = "A1",
+      bottom_color  = c(rgb = "FF000000"),
+      left_color    = c(rgb = "FF000000"),
+      right_color   = c(rgb = "FF000000"),
+      top_color     = c(rgb = "FF000000"),
+      bottom_border = "thin",
+      left_border   = "thin",
+      right_border  = "thin",
+      top_border    = "thin",
+      inner_hgrid   = NULL,
+      inner_hcolor  = NULL,
+      inner_vgrid   = NULL,
+      inner_vcolor  = NULL
+    ) {
+
+      # TODO merge styles and if a style is already present, only add the newly
+      # created border style
+
+      # cc <- wb$worksheets[[sheet]]$sheet_data$cc
+      # df_s <- as.data.frame(lapply(df, function(x) cc$c_s[cc$r %in% x]))
+
+      df <- dims_to_dataframe(dims, fill = TRUE)
+      sheet <- private$get_sheet_index(sheet)
+
+      ### beg border creation
+      full_single <- create_border(
+        top = top_border, top_color = top_color,
+        bottom = bottom_border, bottom_color = bottom_color,
+        left = left_border, left_color = left_color,
+        right = left_border, right_color = right_color
+      )
+
+      top_single <- create_border(
+        top = top_border, top_color = top_color,
+        bottom = inner_vgrid, bottom_color = inner_vcolor,
+        left = left_border, left_color = left_color,
+        right = left_border, right_color = right_color
+      )
+
+      middle_single <- create_border(
+        top = inner_vgrid, top_color = inner_vcolor,
+        bottom = inner_vgrid, bottom_color = inner_vcolor,
+        left = left_border, left_color = left_color,
+        right = left_border, right_color = right_color
+      )
+
+      bottom_single <- create_border(
+        top = inner_vgrid, top_color = inner_vcolor,
+        bottom = bottom_border, bottom_color = bottom_color,
+        left = left_border, left_color = left_color,
+        right = left_border, right_color = right_color
+      )
+
+      left_single <- create_border(
+        top = top_border, top_color = top_color,
+        bottom = bottom_border, bottom_color = bottom_color,
+        left = left_border, left_color = left_color,
+        right = inner_hgrid, right_color = inner_hcolor
+      )
+
+      right_single <- create_border(
+        top = top_border, top_color = top_color,
+        bottom = bottom_border, bottom_color = bottom_color,
+        left = inner_hgrid, left_color = inner_hcolor,
+        right = right_border, right_color = right_color
+      )
+
+      center_single <- create_border(
+        top = top_border, top_color = top_color,
+        bottom = bottom_border, bottom_color = bottom_color,
+        left = inner_hgrid, left_color = inner_hcolor,
+        right = inner_hgrid, right_color = inner_hcolor
+      )
+
+      top_left <- create_border(
+        top = top_border, top_color = top_color,
+        bottom = inner_vgrid, bottom_color = inner_vcolor,
+        left = left_border, left_color = left_color,
+        right = inner_hgrid, right_color = inner_hcolor
+      )
+
+      top_right <- create_border(
+        top = top_border, top_color = top_color,
+        bottom = inner_vgrid, bottom_color = inner_vcolor,
+        left = inner_hgrid, left_color = inner_hcolor,
+        right = left_border, right_color = right_color
+      )
+
+      bottom_left <- create_border(
+        top = inner_vgrid, top_color = inner_vcolor,
+        bottom = bottom_border, bottom_color = bottom_color,
+        left = left_border, left_color = left_color,
+        right = inner_hgrid, right_color = inner_hcolor
+      )
+
+      bottom_right <- create_border(
+        top = inner_vgrid, top_color = inner_vcolor,
+        bottom = bottom_border, bottom_color = bottom_color,
+        left = inner_hgrid, left_color = inner_hcolor,
+        right = left_border, right_color = right_color
+      )
+
+      top_center <- create_border(
+        top = top_border, top_color = top_color,
+        bottom = inner_vgrid, bottom_color = inner_vcolor,
+        left = inner_hgrid, left_color = inner_hcolor,
+        right = inner_hgrid, right_color = inner_hcolor
+      )
+
+      bottom_center <- create_border(
+        top = inner_vgrid, top_color = inner_vcolor,
+        bottom = bottom_border, bottom_color = bottom_color,
+        left = inner_hgrid, left_color = inner_hcolor,
+        right = inner_hgrid, right_color = inner_hcolor
+      )
+
+      middle_left <- create_border(
+        top = inner_vgrid, top_color = inner_vcolor,
+        bottom = inner_vgrid, bottom_color = inner_vcolor,
+        left = left_border, left_color = left_color,
+        right = inner_hgrid, right_color = inner_hcolor
+      )
+
+      middle_right <- create_border(
+        top = inner_vgrid, top_color = inner_vcolor,
+        bottom = inner_vgrid, bottom_color = inner_vcolor,
+        left = inner_hgrid, left_color = inner_hcolor,
+        right = right_border, right_color = right_color
+      )
+
+      inner_cell <- create_border(
+        top = inner_vgrid, top_color = inner_vcolor,
+        bottom = inner_vgrid, bottom_color = inner_vcolor,
+        left = inner_hgrid, left_color = inner_hcolor,
+        right = inner_hgrid, right_color = inner_hcolor
+      )
+      ### end border creation
+
+      #
+      # /* top_single    */
+      # /* middle_single */
+      # /* bottom_single */
+      #
+
+      # /* left_single --- center_single --- right_single */
+
+      #
+      # /* top_left   --- top_center   ---  top_right */
+      # /*  -                                    -    */
+      # /*  -                                    -    */
+      # /*  -                                    -    */
+      # /* left_middle                   right_middle */
+      # /*  -                                    -    */
+      # /*  -                                    -    */
+      # /*  -                                    -    */
+      # /* left_bottom - bottom_center - bottom_right */
+      #
+
+      ## beg cell references
+      if (ncol(df) == 1 && nrow(df) == 1)
+        dim_full_single <- df[1, 1]
+
+      if (ncol(df) == 1 && nrow(df) >= 2) {
+        dim_top_single <- df[1,1]
+        dim_bottom_single <- df[nrow(df), 1]
+        if (nrow(df) >= 3) {
+          mid <- df[,1]
+          dim_middle_single <- mid[!mid %in% c(dim_top_single, dim_bottom_single)]
+        }
+      }
+
+      if (ncol(df) >= 2 && nrow(df) == 1) {
+        dim_left_single <- df[1,1]
+        dim_right_single <- df[1, ncol(df)]
+        if (ncol(df) >= 3) {
+          ctr <- df[1,]
+          dim_center_single <- ctr[!ctr %in% c(dim_left_single, dim_right_single)]
+        }
+      }
+
+      if (ncol(df) >= 2 && nrow(df) >= 2) {
+        dim_top_left     <- df[1, 1]
+        dim_bottom_left  <- df[nrow(df), 1]
+        dim_top_right    <- df[1, ncol(df)]
+        dim_bottom_right <- df[nrow(df), ncol(df)]
+
+        if (nrow(df) >= 3) {
+          top_mid <- df[,1]
+          bottom_mid <- df[,ncol(df)]
+
+          dim_middle_left <- top_mid[!top_mid %in% c(dim_top_left, dim_bottom_left)]
+          dim_middle_right <- bottom_mid[!bottom_mid %in% c(dim_top_right, dim_bottom_right)]
+        }
+
+        if (ncol(df) >= 3) {
+          top_ctr <- df[1,]
+          bottom_ctr <- df[nrow(df),]
+
+          dim_top_center <- top_ctr[!top_ctr %in% c(dim_top_left, dim_top_right)]
+          dim_bottom_center <- bottom_ctr[!bottom_ctr %in% c(dim_bottom_left, dim_bottom_right)]
+        }
+
+        if (ncol(df) > 2 && nrow(df) > 2) {
+          t_row <- 1
+          b_row <- nrow(df)
+          l_row <- 1
+          r_row <- ncol(df)
+          dim_inner_cell <- as.character(unlist(df[c(-t_row, -b_row), c(-l_row, -r_row)]))
+        }
+      }
+      ### end cell references
+
+      # add some random string to the name. if called multiple times, new
+      # styles will be created. We do not look for identical styles, therefor
+      # we might create duplicates, but if a single style changes, the rest of
+      # the workbook remains valid.
+      smp <- paste0(sample(letters, size = 6, replace = TRUE), collapse = "")
+      s <- function(x) paste0(smp, "s", deparse(substitute(x)), seq_along(x))
+      sfull_single <- paste0(smp, "full_single")
+      stop_single <- paste0(smp, "full_single")
+      sbottom_single <- paste0(smp, "bottom_single")
+      smiddle_single <- paste0(smp, "middle_single")
+      sleft_single <- paste0(smp, "left_single")
+      sright_single <- paste0(smp, "right_single")
+      scenter_single <- paste0(smp, "center_single")
+      stop_left <- paste0(smp, "top_left")
+      stop_right <- paste0(smp, "top_right")
+      sbottom_left <- paste0(smp, "bottom_left")
+      sbottom_right <- paste0(smp, "bottom_right")
+      smiddle_left <- paste0(smp, "middle_left")
+      smiddle_right <- paste0(smp, "middle_right")
+      stop_center <- paste0(smp, "top_center")
+      sbottom_center <- paste0(smp, "bottom_center")
+      sinner_cell <- paste0(smp, "inner_cell")
+
+      # ncol == 1
+      if (ncol(df) == 1) {
+
+        # single cell
+        if (nrow(df) == 1) {
+          self$styles_mgr$add(full_single, sfull_single)
+          xf_prev <- get_cell_styles(self, sheet, dims)
+          xf_full_single <- set_border(xf_prev, self$styles_mgr$get_border_id(sfull_single))
+          self$styles_mgr$add(xf_full_single, xf_full_single)
+          set_cell_style(self, sheet, dims, self$styles_mgr$get_xf_id(xf_full_single))
+        }
+
+        # create top & bottom piece
+        if (nrow(df) >= 2) {
+
+          # top single
+          self$styles_mgr$add(top_single, stop_single)
+          xf_prev <- get_cell_styles(self, sheet, dim_top_single)
+          xf_top_single <- set_border(xf_prev, self$styles_mgr$get_border_id(stop_single))
+          self$styles_mgr$add(xf_top_single, xf_top_single)
+          set_cell_style(self, sheet, dim_top_single, self$styles_mgr$get_xf_id(xf_top_single))
+
+          # bottom single
+          self$styles_mgr$add(bottom_single, sbottom_single)
+          xf_prev <- get_cell_styles(self, sheet, dim_bottom_single)
+          xf_bottom_single <- set_border(xf_prev, self$styles_mgr$get_border_id(sbottom_single))
+          self$styles_mgr$add(xf_bottom_single, xf_bottom_single)
+          set_cell_style(self, sheet, dim_bottom_single, self$styles_mgr$get_xf_id(xf_bottom_single))
+        }
+
+        # create middle piece(s)
+        if (nrow(df) >= 3) {
+
+          # middle single
+          self$styles_mgr$add(middle_single, smiddle_single)
+          xf_prev <- get_cell_styles(self, sheet, dim_middle_single)
+          xf_middle_single <- set_border(xf_prev, self$styles_mgr$get_border_id(smiddle_single))
+          self$styles_mgr$add(xf_middle_single, xf_middle_single)
+          set_cell_style(self, sheet, dim_middle_single, self$styles_mgr$get_xf_id(xf_middle_single))
+        }
+
+      }
+
+      # create left and right single row pieces
+      if (ncol(df) >= 2 && nrow(df) == 1) {
+
+        # left single
+        self$styles_mgr$add(left_single, sleft_single)
+        xf_prev <- get_cell_styles(self, sheet, dim_left_single)
+        xf_left_single <- set_border(xf_prev, self$styles_mgr$get_border_id(sleft_single))
+        self$styles_mgr$add(xf_left_single, xf_left_single)
+        set_cell_style(self, sheet, dim_left_single, self$styles_mgr$get_xf_id(xf_left_single))
+
+        # right single
+        self$styles_mgr$add(right_single, sright_single)
+        xf_prev <- get_cell_styles(self, sheet, dim_right_single)
+        xf_right_single <- set_border(xf_prev, self$styles_mgr$get_border_id(sright_single))
+        self$styles_mgr$add(xf_right_single, xf_right_single)
+        set_cell_style(self, sheet, dim_right_single, self$styles_mgr$get_xf_id(xf_right_single))
+
+        # add single center piece(s)
+        if (ncol(df) >= 3) {
+
+          # center single
+          self$styles_mgr$add(center_single, scenter_single)
+          xf_prev <- get_cell_styles(self, sheet, dim_center_single)
+          xf_center_single <- set_border(xf_prev, self$styles_mgr$get_border_id(scenter_single))
+          self$styles_mgr$add(xf_center_single, xf_center_single)
+          set_cell_style(self, sheet, dim_center_single, self$styles_mgr$get_xf_id(xf_center_single))
+        }
+
+      }
+
+      # create left & right - top & bottom corners pieces
+      if (ncol(df) >= 2 && nrow(df) >= 2) {
+
+        # top left
+        self$styles_mgr$add(top_left, stop_left)
+        xf_prev <- get_cell_styles(self, sheet, dim_top_left)
+        xf_top_left <- set_border(xf_prev, self$styles_mgr$get_border_id(stop_left))
+        self$styles_mgr$add(xf_top_left, xf_top_left)
+        set_cell_style(self, sheet, dim_top_left, self$styles_mgr$get_xf_id(xf_top_left))
+
+        # top right
+        self$styles_mgr$add(top_right, stop_right)
+        xf_prev <- get_cell_styles(self, sheet, dim_top_right)
+        xf_top_right <- set_border(xf_prev, self$styles_mgr$get_border_id(stop_right))
+        self$styles_mgr$add(xf_top_right, xf_top_right)
+        set_cell_style(self, sheet, dim_top_right, self$styles_mgr$get_xf_id(xf_top_right))
+
+        # bottom left
+        self$styles_mgr$add(bottom_left, sbottom_left)
+        xf_prev <- get_cell_styles(self, sheet, dim_bottom_left)
+        xf_bottom_left <- set_border(xf_prev, self$styles_mgr$get_border_id(sbottom_left))
+        self$styles_mgr$add(xf_bottom_left, xf_bottom_left)
+        set_cell_style(self, sheet, dim_bottom_left, self$styles_mgr$get_xf_id(xf_bottom_left))
+
+        # bottom right
+        self$styles_mgr$add(bottom_right, sbottom_right)
+        xf_prev <- get_cell_styles(self, sheet, dim_bottom_right)
+        xf_bottom_right <- set_border(xf_prev, self$styles_mgr$get_border_id(sbottom_right))
+        self$styles_mgr$add(xf_bottom_right, xf_bottom_right)
+        set_cell_style(self, sheet, dim_bottom_right, self$styles_mgr$get_xf_id(xf_bottom_right))
+      }
+
+      # create left and right middle pieces
+      if (ncol(df) >= 2 && nrow(df) >= 3) {
+
+        # middle left
+        self$styles_mgr$add(middle_left, smiddle_left)
+        xf_prev <- get_cell_styles(self, sheet, dim_middle_left)
+        xf_middle_left <- set_border(xf_prev, self$styles_mgr$get_border_id(smiddle_left))
+        self$styles_mgr$add(xf_middle_left, xf_middle_left)
+        set_cell_style(self, sheet, dim_middle_left, self$styles_mgr$get_xf_id(xf_middle_left))
+
+        # middle right
+        self$styles_mgr$add(middle_right, smiddle_right)
+        xf_prev <- get_cell_styles(self, sheet, dim_middle_right)
+        xf_middle_right <- set_border(xf_prev, self$styles_mgr$get_border_id(smiddle_right))
+        self$styles_mgr$add(xf_middle_right, xf_middle_right)
+        set_cell_style(self, sheet, dim_middle_right, self$styles_mgr$get_xf_id(xf_middle_right))
+      }
+
+      # create top and bottom center pieces
+      if (ncol(df) >= 3 & nrow(df) >= 2) {
+
+        # top center
+        self$styles_mgr$add(top_center, stop_center)
+        xf_prev <- get_cell_styles(self, sheet, dim_top_center)
+        xf_top_center <- set_border(xf_prev, self$styles_mgr$get_border_id(stop_center))
+        self$styles_mgr$add(xf_top_center, xf_top_center)
+        set_cell_style(self, sheet, dim_top_center, self$styles_mgr$get_xf_id(xf_top_center))
+
+        # bottom center
+        self$styles_mgr$add(bottom_center, sbottom_center)
+        xf_prev <- get_cell_styles(self, sheet, dim_bottom_center)
+        xf_bottom_center <- set_border(xf_prev, self$styles_mgr$get_border_id(sbottom_center))
+        self$styles_mgr$add(xf_bottom_center, xf_bottom_center)
+        set_cell_style(self, sheet, dim_bottom_center, self$styles_mgr$get_xf_id(xf_bottom_center))
+      }
+
+      if (nrow(df) > 2 && ncol(df) > 2) {
+
+        # inner cells
+        self$styles_mgr$add(inner_cell, sinner_cell)
+        xf_prev <- get_cell_styles(self, sheet, dim_inner_cell)
+        xf_inner_cell <- set_border(xf_prev, self$styles_mgr$get_border_id(sinner_cell))
+        self$styles_mgr$add(xf_inner_cell, xf_inner_cell)
+        set_cell_style(self, sheet, dim_inner_cell, self$styles_mgr$get_xf_id(xf_inner_cell))
+      }
+
+      return(self)
+    },
+
+    #' @description provide simple fill function
+    #' @param sheet the worksheet
+    #' @param dims the cell range
+    #' @param color the colors to apply, e.g. yellow: c(rgb = "FFFFFF00")
+    #' @param pattern various default "none" but others are possible:
+    #'  "solid", "mediumGray", "darkGray", "lightGray", "darkHorizontal",
+    #'  "darkVertical", "darkDown", "darkUp", "darkGrid", "darkTrellis",
+    #'  "lightHorizontal", "lightVertical", "lightDown", "lightUp", "lightGrid",
+    #'  "lightTrellis", "gray125", "gray0625"
+    #' @param gradient_fill a gradient fill xml pattern.
+    #' @param every_nth_col which col should be filled
+    #' @param every_nth_row which row should be filled
+    #' @examples
+    #'  # example from the gradient fill manual page
+    #'  gradient_fill <- "<gradientFill degree=\"90\">
+    #'    <stop position=\"0\"><color rgb=\"FF92D050\"/></stop>
+    #'    <stop position=\"1\"><color rgb=\"FF0070C0\"/></stop>
+    #'   </gradientFill>"
+    #' @return The `wbWorksheetObject`, invisibly
+    add_fill = function(
+      sheet,
+      dims,
+      color = "",
+      pattern = "solid",
+      gradient_fill = "",
+      every_nth_col = 1,
+      every_nth_row = 1
+    ) {
+
+      new_fill <- create_fill(
+        gradientFill = gradient_fill,
+        patternType = pattern,
+        fgColor = color
+      )
+
+      smp <- paste0(sample(letters, size = 6, replace = TRUE), collapse = "")
+      snew_fill <- paste0(smp, "new_fill")
+      sxf_new_fill <- paste0(smp, "xf_new_fill")
+
+
+      self$styles_mgr$add(new_fill, snew_fill)
+
+      # dim in dataframe can contain various styles. go cell by cell.
+      did <- dims_to_dataframe(dims, fill = TRUE)
+      # select a few cols and rows to fill
+      cols <- (seq_len(ncol(did)) %% every_nth_col) == 0
+      rows <- (seq_len(nrow(did)) %% every_nth_row) == 0
+
+      dims <- unname(unlist(did[rows, cols, drop = FALSE]))
+
+      for (dim in dims) {
+        sxf_new_fill_x <- paste0(sxf_new_fill, which(dims %in% dim))
+        xf_prev <- get_cell_styles(self, sheet, dim)
+        xf_new_fill <- set_fill(xf_prev, self$styles_mgr$get_fill_id(snew_fill))
+        self$styles_mgr$add(xf_new_fill, xf_new_fill)
+        s_id <- self$styles_mgr$get_xf_id(xf_new_fill)
+        set_cell_style(self, sheet, dim, s_id)
+      }
+
+      return(self)
     }
+
   ),
 
   ## private ----
@@ -4124,8 +4614,8 @@ wbWorkbook <- R6::R6Class(
 
     ### methods ----
     deep_clone = function(name, value) {
-      #' Deep cloning method for workbooks.  This method also accesses
-      #' `$clone(deep = TRUE)` methods for `R6` fields.
+      # Deep cloning method for workbooks.  This method also accesses
+      # `$clone(deep = TRUE)` methods for `R6` fields.
       if (R6::is.R6(value)) {
         value <- value$clone(deep = TRUE)
       } else if (is.list(value)) {